Restructure project into src package
