+++ src/sentience/core.py
