+++ src/sentience/core_old.py
